# docs/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "version": "1.0.0",
    "title": "Phonebook App",
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "description": "Phonebook app allowes to store user's contacts in the database, edit them, and access to them from anywhere on Earth having internet connection."
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    },
    {
      "url": "https://example.com/api/v1"
    }
  ],
  "tags": [
    {
      "name": "Contacts",
      "description": "Operations with contacts."
    },
    {
      "name": "Auth",
      "description": "Auth operations."
    }
  ],
  "paths": {
    "/contacts": {
      "get": {
        "tags": [
          "Contacts"
        ],
        "summary": "Get a list of the user's contacts",
        "operationId": "getAllContacts",
        "description": "Get **an array** of objects representing **the contacts** created by the user. </br> Use **filters** as query parameters to narrow down the server response. </br> **Sort** response by the specific parameter in ascending or descending order. </br> Use **pagination** by passing the page number and number of contacts per page in the query parameters.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer"
            },
            "example": 1
          },
          {
            "in": "query",
            "name": "perPage",
            "schema": {
              "type": "integer",
              "example": 10
            }
          },
          {
            "in": "query",
            "name": "sortBy",
            "schema": {
              "type": "string",
              "example": "name"
            },
            "description": "All contacts's fields can be mentioned. Default value - _id"
          },
          {
            "in": "query",
            "name": "sortOrder",
            "schema": {
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ],
              "example": "asc"
            }
          },
          {
            "in": "query",
            "name": "contactType",
            "schema": {
              "type": "string",
              "enum": [
                "work",
                "home",
                "personal"
              ],
              "example": "home"
            }
          },
          {
            "in": "query",
            "name": "isFavourite",
            "schema": {
              "type": "boolean",
              "example": true
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/getAllContacts-200"
          },
          "400": {
            "$ref": "#/components/responses/getAllContacts-400"
          },
          "401": {
            "$ref": "#/components/responses/401"
          }
        }
      },
      "post": {
        "tags": [
          "Contacts"
        ],
        "summary": "Create a new contact",
        "operationId": "createContact",
        "description": "**Create** a new contact. Request contains **a body** as a json object with data of the new contact.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "phoneNumber",
                  "contactType"
                ],
                "properties": {
                  "name": {
                    "description": "The contact's name",
                    "type": "string",
                    "minLength": 3,
                    "maxLength": 20,
                    "example": "John Snow"
                  },
                  "phoneNumber": {
                    "description": "The phone number of the contact",
                    "type": "string",
                    "example": "+380501234567"
                  },
                  "email": {
                    "description": "Must be a valid email",
                    "type": "string",
                    "example": "john.snow@winterfell.com"
                  },
                  "isFavourite": {
                    "description": "Is the contact a favourite?",
                    "type": "boolean",
                    "example": false
                  },
                  "contactType": {
                    "description": "Contact group by area of their use",
                    "type": "string",
                    "enum": [
                      "work",
                      "home",
                      "personal"
                    ],
                    "example": "personal"
                  },
                  "photo": {
                    "description": "Link to the contact's avatar image",
                    "type": "string",
                    "example": "https://res.cloudinary.com/dbkdexyl9/image/upload/v1745741839/swbjzex5zv6lsncc4d3e.jpg"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "$ref": "#/components/responses/createContact-201"
          },
          "400": {
            "$ref": "#/components/responses/createContact-400"
          },
          "401": {
            "$ref": "#/components/responses/401"
          }
        }
      }
    },
    "/contact/{contactId}": {
      "get": {
        "tags": [
          "Contacts"
        ],
        "summary": "Get a contact by its id",
        "operationId": "getContactById",
        "description": "**Get** an object representing the **contact** by its **id** passed as an url parameter [/contacts/:contactId](/contacts/:contactId). </br> Parameter **contactId** must be a valid id: 24 hexadecimal symbols.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "contactId",
            "required": true,
            "schema": {
              "type": "string",
              "example": "67f6381e907fe15ef69f1df0"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/getContactById-200"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "404": {
            "$ref": "#/components/responses/getContactById-404"
          }
        }
      },
      "delete": {
        "tags": [
          "Contacts"
        ],
        "summary": "Delete the user's contact by its id",
        "operationId": "deleteContact",
        "description": "**Delete** an object representing the **contact** by its **id** passed as an url parameter [/contacts/:contactId](/contacts/:contactId). </br> Parameter **contactId** must be a valid id: 24 hexadecimal symbols.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "contactId",
            "required": true,
            "schema": {
              "type": "string",
              "example": "67f6381e907fe15ef69f1df0"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "The contact has deleted successfully."
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "404": {
            "$ref": "#/components/responses/getContactById-404"
          }
        }
      },
      "put": {
        "tags": [
          "Contacts"
        ],
        "summary": "Upsert a contact.",
        "operationId": "upsertContact",
        "description": "**Update** existing contact or **create** the new one if it's absent in the database. Request contains **a body** as a **json** object with contact update data or **multipart/formdata** body that can contain avatar image",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "contactId",
            "required": true,
            "schema": {
              "type": "string",
              "example": "67f6381e907fe15ef69f1df0"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "phoneNumber",
                  "contactType"
                ],
                "properties": {
                  "name": {
                    "description": "The contact's name",
                    "type": "string",
                    "minLength": 3,
                    "maxLength": 20,
                    "example": "John Snow"
                  },
                  "phoneNumber": {
                    "description": "The phone number of the contact",
                    "type": "string",
                    "example": "+380501234567"
                  },
                  "email": {
                    "description": "Must be a valid email",
                    "type": "string",
                    "example": "john.snow@winterfell.com"
                  },
                  "isFavourite": {
                    "description": "Is the contact a favourite?",
                    "type": "boolean",
                    "example": false
                  },
                  "contactType": {
                    "description": "Contact group by area of their use",
                    "type": "string",
                    "enum": [
                      "work",
                      "home",
                      "personal"
                    ],
                    "example": "personal"
                  },
                  "photo": {
                    "description": "Link to the contact's avatar image",
                    "type": "string",
                    "example": "https://res.cloudinary.com/dbkdexyl9/image/upload/v1745741839/swbjzex5zv6lsncc4d3e.jpg"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/putContact-200"
          },
          "201": {
            "$ref": "#/components/responses/createContact-201"
          },
          "400": {
            "$ref": "#/components/responses/createContact-400"
          },
          "401": {
            "$ref": "#/components/responses/401"
          }
        }
      },
      "patch": {
        "tags": [
          "Contacts"
        ],
        "summary": "Patch existing contact",
        "operationId": "patchContact",
        "description": "**Patch** existing contact by its id. Request contains **a body** as a **json** object with contact update data or **multipart/formdata** body that can contain avatar image",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "contactId",
            "required": true,
            "schema": {
              "type": "string",
              "example": "67f6381e907fe15ef69f1df0"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "description": "The contact's name",
                    "type": "string",
                    "minLength": 3,
                    "maxLength": 20,
                    "example": "John Snow"
                  },
                  "phoneNumber": {
                    "description": "The phone number of the contact",
                    "type": "string",
                    "example": "+380501234567"
                  },
                  "email": {
                    "description": "Must be a valid email",
                    "type": "string",
                    "example": "john.snow@winterfell.com"
                  },
                  "isFavourite": {
                    "description": "Is the contact a favourite?",
                    "type": "boolean",
                    "example": false
                  },
                  "contactType": {
                    "description": "Contact group by area of their use",
                    "type": "string",
                    "enum": [
                      "work",
                      "home",
                      "personal"
                    ],
                    "example": "personal"
                  },
                  "photo": {
                    "description": "Link to the contact's avatar image",
                    "type": "string",
                    "example": "https://res.cloudinary.com/dbkdexyl9/image/upload/v1745741839/swbjzex5zv6lsncc4d3e.jpg"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/patchContact-200"
          },
          "400": {
            "$ref": "#/components/responses/createContact-400"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "404": {
            "$ref": "#/components/responses/getContactById-404"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "schemas": {
      "contact": {
        "type": "object",
        "required": [
          "_id",
          "userId",
          "name",
          "phoneNumber",
          "contactType"
        ],
        "properties": {
          "_id": {
            "description": "Contact's id",
            "type": "string",
            "example": "67f6381e907fe15ef69f1df0"
          },
          "userId": {
            "description": "The id of the user who created this contact",
            "type": "string",
            "example": "67f6364eeba7c0aedb007d73"
          },
          "name": {
            "description": "The contact's name",
            "type": "string",
            "minLength": 3,
            "maxLength": 20,
            "example": "John Snow"
          },
          "phoneNumber": {
            "description": "The phone number of the contact",
            "type": "string",
            "example": "+380501234567"
          },
          "email": {
            "description": "Must be a valid email",
            "type": "string",
            "example": "john.snow@winterfell.com"
          },
          "isFavourite": {
            "description": "Is the contact a favourite?",
            "type": "boolean",
            "example": false
          },
          "contactType": {
            "description": "Contact group by area of their use",
            "type": "string",
            "enum": [
              "work",
              "home",
              "personal"
            ],
            "example": "personal"
          },
          "photo": {
            "description": "Link to the contact's avatar image",
            "type": "string",
            "example": "https://res.cloudinary.com/dbkdexyl9/image/upload/v1745741839/swbjzex5zv6lsncc4d3e.jpg"
          }
        }
      }
    },
    "responses": {
      "401": {
        "description": "Unauthorized request",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "status",
                "message",
                "data"
              ],
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 401
                },
                "message": {
                  "type": "string",
                  "example": "Unauthorized"
                },
                "data": {
                  "type": "object",
                  "required": [
                    "message"
                  ],
                  "properties": {
                    "message": {
                      "type": "string",
                      "enum": [
                        "Session not found",
                        "User not found",
                        "Auth header should be of type Bearer",
                        "Please provide Authorization header",
                        "Access token expired",
                        "Refresh token expired",
                        "Token is expired or invalid."
                      ],
                      "example": "Access token expired"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "getAllContacts-200": {
        "description": "Found all contacts using pagination, filter and sort query parameters.",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "status",
                "message",
                "data"
              ],
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 200
                },
                "message": {
                  "type": "string",
                  "example": "Successfully found the contacts!"
                },
                "data": {
                  "type": "object",
                  "required": [
                    "data",
                    "page",
                    "perPage",
                    "totalItems",
                    "totalPages",
                    "hasPreviousPage",
                    "hasNextPage"
                  ],
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/contact"
                      }
                    },
                    "page": {
                      "type": "integer",
                      "minimum": 1,
                      "example": 1
                    },
                    "perPage": {
                      "type": "integer",
                      "minimum": 1,
                      "example": 10
                    },
                    "totalItems": {
                      "type": "integer",
                      "minimum": 0,
                      "example": 50
                    },
                    "totalPages": {
                      "type": "integer",
                      "minimum": 0,
                      "example": 5
                    },
                    "hasPreviousPage": {
                      "type": "boolean",
                      "example": false
                    },
                    "hasNextPage": {
                      "type": "boolean",
                      "example": true
                    }
                  }
                }
              }
            }
          }
        }
      },
      "getAllContacts-400": {
        "description": "Wrong query parameters",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "status",
                "message",
                "data"
              ],
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 400
                },
                "message": {
                  "type": "string",
                  "example": "Bad request"
                },
                "data": {
                  "type": "object",
                  "required": [
                    "message"
                  ],
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "The current page (5) must be in the following range [1, 4]"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "createContact-201": {
        "description": "Successfully created a contact! Response contains the json object with data of the new contact.",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "status",
                "message",
                "data"
              ],
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 201
                },
                "message": {
                  "type": "string",
                  "example": "Successfully created a contact!"
                },
                "data": {
                  "type": "object",
                  "$ref": "#/components/schemas/contact"
                }
              }
            }
          }
        }
      },
      "createContact-400": {
        "description": "Wrong body data",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "status",
                "message",
                "data"
              ],
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 400
                },
                "message": {
                  "type": "string",
                  "example": "Bad request"
                },
                "data": {
                  "type": "object",
                  "required": [
                    "message"
                  ],
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "\"contactType\" must be one of [work, home, personal]"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "getContactById-200": {
        "description": "Successfully found the contact by its id.",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "status",
                "message",
                "data"
              ],
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 200
                },
                "message": {
                  "type": "string",
                  "example": "Successfully found a contact with id: 67f6381e907fe15ef69f1df0"
                },
                "data": {
                  "type": "object",
                  "$ref": "#/components/schemas/contact"
                }
              }
            }
          }
        }
      },
      "getContactById-404": {
        "description": "Contact is absent in the database",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "status",
                "message",
                "data"
              ],
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 404
                },
                "message": {
                  "type": "string",
                  "example": "Contact not found!"
                },
                "data": {
                  "type": "object",
                  "required": [
                    "message"
                  ],
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Contact not found!"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "putContact-200": {
        "description": "Successfully updated a contact! Response contains the json object with data of the updated contact.",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "status",
                "message",
                "data"
              ],
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 200
                },
                "message": {
                  "type": "string",
                  "example": "Successfully updated a contact!"
                },
                "data": {
                  "type": "object",
                  "$ref": "#/components/schemas/contact"
                }
              }
            }
          }
        }
      },
      "patchContact-200": {
        "description": "Successfully patched a contact! Response contains the json object with data of the patched contact.",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "status",
                "message",
                "data"
              ],
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 200
                },
                "message": {
                  "type": "string",
                  "example": "Successfully patched a contact!"
                },
                "data": {
                  "type": "object",
                  "$ref": "#/components/schemas/contact"
                }
              }
            }
          }
        }
      }
    }
  }
}